refactor(filter): Pull out pattern logic
